requirements.txt
aniso8601==8.0.0
cffi==1.13.2
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.3.0
mysql-connector-python==8.0.18
pipdeptree==0.13.2
protobuf==3.11.2
pycparser==2.19
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.3
six==1.13.0
Werkzeug==0.16.0
